backend/requirements/base.txt
# Base dependencies required for all environments
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
python-multipart>=0.0.6

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# HTTP client
httpx>=0.25.0

# Utilities
email-validator>=2.1.0
psutil>=5.9.0

# Redis
redis>=5.0.0
aioredis>=2.0.0

# Rate limiting
slowapi>=0.1.9